Fix check of return value of aio_return()
The man page was misleading and I must have had a bad moment mechanically
coding what the man page says without thinking about it.
So I seem to have assumed that aio_return() might return EINVAL which can't be
the case. In fact the error is returned in errno which isn't mentioned by the
Linux man page for aio_return.