RelayReader: lengthen read timeout
That was necessary to proxy an usb-stick here. Otherwise the first
bulk message after the config setup ended up in a timeout.

This isn't a very good solution, more a workaround.

A better solution would be if someone takes the time to read the libusb
documentation and rewrites the stuff in DeviceProxy_LibUSB.cpp.

For me it looks like if a timeout occurs in DeviceProxy_LibUSB,
usbproxy falls apart. I haven't looked deeper at the stuff.