USB Mass Storages fix with LibUSB
Change receive_data in DeviceProxy_LibUSB.cpp to fix problems with USB Mass Storages.

I've read everywhere that everyone has problems with USB Mass Storages not being recognized. 
The cause is that after device setup, the host sends an INQUIRY request (0x12) to get device information (http://www.usb.org/developers/docs/devclass_docs/usbmass-ufi10.pdf, page 19). The device takes a while before responding and starting to send packets to the host, meanwhile libUSB timeout triggers and the operation fails. 

I don't know if what I said before is correct, I'm really newbie with USB Standard and I've only sniffed packets trying to solve the problem; but the working solution (tested with various USB Mass Storages with various file systems) is to add a simple do-while that waits the response from the mass storage. 

We make 5 attempts and then break the loop because the problem is another, and not the INQUIRY command, so we avoid infinite loops.

The timeout = 0 is needed because in my tests, sometimes the device goes in timeout while reading/writing blocks, but it needs further testing (it's working for me so I've left it there).

I don't know if the same stuff is needed in send_data also, as I said it's working for me so if you want to add it just do it.