Fix GET_STATUS control requests not being forwarded
If bRequest = 0, the request would be ignored. As a result, all USB GET_STATUS
device requests where ignored.

Instead of using bRequest as "return value" to decide if a setup packet should
be generated or not in RelayReader, use the actual return value of the function
to make this decission.

Signed-off-by: Sam Van Den Berge <sam.van.den.berge@telenet.be>